Supports PUT and POST
Introduces verbs to the Route annotation -- GET is the default. The route handler will fetch the request body and JSON decode it if the method is either PUT or POST. The proper path variables are passed to the route handler. One route name can be used by more than one route, as long as the verbs are different.

The route handlers for a resource are called on an instance of the resource. To be generic, this requires that all resources have constructors that take no arguments. Otherwise, all methods would have to be static, which we do not want.